// src/lib/Enums/SensorTypeEnum.ts
enum SensorTypeEnum {
    SENSOR_TYPE_CURRENT = "current",
    SENSOR_TYPE_TEMPERATURE = "temperature",
    SENSOR_TYPE_HUMIDITY = "humidity",
    SENSOR_TYPE_PRESSURE = "pressure",
    SENSOR_TYPE_WIND_SPEED = "wind_speed",
    SENSOR_TYPE_WIND_DIRECTION = "wind_direction",
    SENSOR_TYPE_RAIN = "rain",
    SENSOR_TYPE_UV = "uv",
    SENSOR_TYPE_SOLAR_RADIATION = "solar_radiation",
    SENSOR_TYPE_UNKNOWN = "unknown",
}

export default SensorTypeEnum;